requirements.txt
numpy<2.0
faiss-cpu==1.8.0.post1
streamlit==1.48.1
openai==1.101.0
notion-client==2.4.0
pandas==2.3.2
